requirements.txt
keboola.component==1.3.5
keboola.utils
keboola.http-client
https://bitbucket.org/kds_consulting_team/keboola-python-util-lib/get/0.5.0.zip#egg=kbc
mock
freezegun
ibm-cos-sdk
ibm-watson==5.3.0
botocore
oauth2client
pandas
openpyxl
